Add all graphs

+++ task-12-1/Graph/inc/igraph.h
@@ -12,7 +12,7 @@ struct IGraph {
   // Добавление ребра от from к to.
   virtual void AddEdge(size_t from, size_t to) = 0;
 
-  virtual int VerticesCount() const = 0;
+  virtual size_t VerticesCount() const = 0;
 
   virtual std::vector<size_t> GetNextVertices(size_t vertex) const = 0;
   virtual std::vector<size_t> GetPrevVertices(size_t vertex) const = 0;
